feat: [api-handlers-decomposition-005] - Preserve model catalog and invocation HTTP behavior
