+++ docs/internal/processes/development-guide-relevant-files.md
@@ -136,8 +136,8 @@ This inventory records the checked-in files and directories that the maintainer 
 | `.github/workflows/ci.yml` | Repository CI contract | Contributor docs should name this workflow and mirror its root-level validation commands and stated non-deployment scope; when a required lane needs actionable diagnostics, prefer separate workflow steps that each call the existing repo-owned root commands such as `make ui-build`, `make build`, `make lint`, `make api-smoke`, `make ui-test`, `make ui-integration-test`, and `make test-coverage-go` instead of hiding the whole lane behind one opaque composite step. Keep Playwright Chromium installation immediately before the browser-backed `make ui-integration-test` step so CI and local review flows exercise the same dedicated lane. The backend coverage job should continue to call `make test-coverage-go` so CI and local runs share the same repo-owned `cmd/gocoveragecheck` policy surface. |
 | `internal/testpath/artifact_contract.go` | Enforced root artifact contract | Root-level factory artifact additions, removals, and redirect stubs are test-enforced here and must stay synchronized with the inventory doc. |
 | `factory/` | Maintainer workflow surface | Contains checked-in operator guidance and active inbox directories that the development guide may reference for workflow-related tasks. |
-| `pkg/api/handlers.go` plus sibling `pkg/api/handlers_*.go` files | Handwritten API boundary | Keep HTTP handler entrypoints, raw-JSON decode or validation helpers, response projection helpers, and shared API response writers in the same package but split across sibling `handlers_*.go` files when size cleanup is needed; shared JSON/SSE writers and `errorFamilyForStatus` live at the top of `handlers_decode.go` beside request-body decode helpers and canonical work-request validation translation; factory validation, current-factory read/save, factory-session open/close/list, and workstation prompt-template routes live in `handlers_factory.go` with `loadCurrentFactory`, `writeCurrentFactoryError`, and prompt-template projection helpers colocated there; submit and work-request upsert routes (`SubmitWork`, `SubmitWorkBySessionId`, `UpsertWorkRequest`, `UpsertWorkRequestBySessionId`) live in `handlers_work.go` with `upsertWorkRequestResponse` and trace-stabilization helpers colocated there. The root `pkg/api` package has a 15-file cap—when adding a new `handlers_*.go` owner file, consolidate or move an existing root source/test file first (for example colocate submit staged-file handlers with `submit_work_items.go`, merge retired `handlers_common.go` into `handlers_decode.go`, or move `workstation_request_projection.go` into `pkg/api/workstationprojection/` with a thin `BuildFactoryWorldWorkstationRequestProjectionSlice` delegate in `handlers.go`). Generated request structs still do not reject unknown keys by themselves, so retired work-request alias and trace-conflict policy should continue routing through `pkg/factory/requests.ValidateCanonicalWorkRequestJSON(...)` via `validateCanonicalWorkRequestJSONForAPI` in `handlers_decode.go` and only be translated into API-owned error wording rather than reimplemented locally. |
-| `pkg/api/server_*_test.go` plus `pkg/api/server_test_helpers_test.go` | API route regression ownership | When API route tests cross the backend size gate, split them into same-package sibling `server_*_test.go` files by boundary concern such as submit, upsert, factory, provider-session, work-query, and dashboard/event coverage, and keep shared request builders or response assertions in `server_test_helpers_test.go` so route behavior stays observable without recreating one giant test owner file. |
+| `pkg/api/handlers.go` plus sibling `pkg/api/handlers_*.go` files | Handwritten API boundary | Keep HTTP handler entrypoints, raw-JSON decode or validation helpers, response projection helpers, and shared API response writers in the same package but split across sibling `handlers_*.go` files when size cleanup is needed; shared JSON/SSE writers and `errorFamilyForStatus` live at the top of `handlers_decode.go` beside request-body decode helpers and canonical work-request validation translation; factory validation, current-factory read/save, factory-session open/close/list, and workstation prompt-template routes live in `handlers_factory.go` with `loadCurrentFactory`, `writeCurrentFactoryError`, and prompt-template projection helpers colocated there; submit and work-request upsert routes (`SubmitWork`, `SubmitWorkBySessionId`, `UpsertWorkRequest`, `UpsertWorkRequestBySessionId`) live in `handlers_work.go` with `upsertWorkRequestResponse` and trace-stabilization helpers colocated there; model catalog and invocation routes (`ListModels`, `GetModel`, `InvokeModel`, `PullModel`) live in `handlers_models.go` with `generatedResolvedModelInvocationBindings` and streaming/audio response mapping colocated there. The root `pkg/api` package has a 15-file cap—when adding a new `handlers_*.go` owner file, consolidate or move an existing root source/test file first (for example colocate submit staged-file handlers with `submit_work_items.go`, merge retired `handlers_common.go` into `handlers_decode.go`, or move `workstation_request_projection.go` into `pkg/api/workstationprojection/` with a thin `BuildFactoryWorldWorkstationRequestProjectionSlice` delegate in `handlers.go`). Generated request structs still do not reject unknown keys by themselves, so retired work-request alias and trace-conflict policy should continue routing through `pkg/factory/requests.ValidateCanonicalWorkRequestJSON(...)` via `validateCanonicalWorkRequestJSONForAPI` in `handlers_decode.go` and only be translated into API-owned error wording rather than reimplemented locally. |
+| `pkg/api/server_*_test.go` plus `pkg/api/server_test_helpers_test.go` | API route regression ownership | When API route tests cross the backend size gate, split them into same-package sibling `server_*_test.go` files by boundary concern such as submit, upsert, factory, provider-session, work-query, and dashboard/event coverage, and keep shared request builders or response assertions in `server_test_helpers_test.go` so route behavior stays observable without recreating one giant test owner file. When the root `pkg/api` file-count cap blocks a new `handlers_*.go` sibling, merge a related undersized `server_*_test.go` into another same-package test owner (for example colocate upsert regressions with `server_factory_test.go`) before adding the handler file. |
 | `api/codegen_config/client.yaml` | Generated Go client config | The checked-in Go client for backend functional tests is generated from this config and must stay in lockstep with `api/openapi.yaml`, `pkg/generatedclient/`, and `make api-smoke`. |
 | `api/openapi-main.yaml` | Authored OpenAPI source of truth | Public API behavior changes should start here, including the advertised local `servers` URL, then refresh the bundled spec and generated Go/UI artifacts together so handwritten route seams and generated parameter binding stay aligned. |
 | `pkg/generatedclient/` | Generated Go API client package | Backend functional tests that need typed client coverage should import this package instead of rebuilding `http.Get` or `http.Post` helpers for OpenAPI-owned routes. |
